feat: first crack
