+++ cli/Cargo.toml
@@ -11,6 +11,10 @@ license.workspace = true
 repository.workspace = true
 description = "Provides the deno executable"
 
+[lib]
+path = "lib.rs"
+name = "deno"
+
 [[bin]]
 name = "deno"
 path = "main.rs"
